web_course_work/full_stack/project/requirements.txt
django-scheduler
werkzeug==0.12.2
Flask==0.12.2
pymongo==3.5.1
beautifulsoup4==4.6.0
passlib == 1.7.1
requests == 2.18.4
